requirements.txt
django
djangorestframework
python-dotenv
spotipy
openai
